+++ frontend/src/app/riesgos/scenarios/ecuador/7_lahardamage.ts
@@ -151,7 +151,8 @@ export const laharUpdatedExposureProps: VectorLayerProperties = {
                             color: getColor(4),
                             text: `mostly_collapsed_damage_lahar`
                         }],
-                        height: 90
+                        height: 90,
+                        width: 330
                     }
                 };
             },